+++ test/Resolver/EmptyIpLocationResolverTest.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace ShlinkioTest\Shlink\IpGeolocation\Resolver;
 
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Shlinkio\Shlink\IpGeolocation\Model\Location;
 use Shlinkio\Shlink\IpGeolocation\Resolver\EmptyIpLocationResolver;
@@ -20,16 +22,13 @@ class EmptyIpLocationResolverTest extends TestCase
         $this->resolver = new EmptyIpLocationResolver();
     }
 
-    /**
-     * @test
-     * @dataProvider provideEmptyResponses
-     */
+    #[Test, DataProvider('provideEmptyResponses')]
     public function alwaysReturnsAnEmptyLocation(string $ipAddress): void
     {
         self::assertEquals(Location::emptyInstance(), $this->resolver->resolveIpLocation($ipAddress));
     }
 
-    public function provideEmptyResponses(): array
+    public static function provideEmptyResponses(): array
     {
         return map(range(0, 5), fn () => ['foobar']);
     }